+++ .eslintrc.cjs
@@ -1,4 +1,4 @@
-export default {
+exports.default = {
   env: { browser: true, es2020: true },
   extends: [
     'eslint:recommended',
