+++ src/hooks/useWindowSize.ts
@@ -4,8 +4,9 @@ import { useEffect, useState } from 'react';
  *
  * @param minWidth Accepts a minWidth (number) and returns wether the size of screen is equal or less than that number.
  *
- * @returns
+ * @returns boolean
  */
+
 export default function useWindowSize(minWidth: number) {
   const [state, setState] = useState({
     windowWidth: window.innerWidth,
